Can now mark up data with corrections to its validity
