+++ src/main/java/org/tahomarobotics/scouting/scoutingserver/controller/TabController.java
@@ -12,6 +12,7 @@ import javafx.stage.FileChooser;
 import javafx.util.Pair;
 import org.json.JSONArray;
 import org.tahomarobotics.scouting.scoutingserver.Constants;
+import org.tahomarobotics.scouting.scoutingserver.DataValidator;
 import org.tahomarobotics.scouting.scoutingserver.DatabaseManager;
 import org.tahomarobotics.scouting.scoutingserver.ScoutingServer;
 import org.tahomarobotics.scouting.scoutingserver.util.APIUtil;
@@ -20,6 +21,7 @@ import org.tahomarobotics.scouting.scoutingserver.util.MatchRecordComparator;
 import org.tahomarobotics.scouting.scoutingserver.util.SpreadsheetUtil;
 
 import java.io.File;
+import java.io.FilterOutputStream;
 import java.io.IOException;
 import java.util.*;
 import java.util.function.Predicate;
@@ -105,108 +107,12 @@ public class TabController {
     }
     @FXML
     public void validateData(ActionEvent event) {
-            String eventCode = "2024week0";
-        try {
-            JSONArray eventMatches = APIUtil.get("/event/" + eventCode + "/matches");//returns array of json objects each representing a match
-            List<Object> rawList =  eventMatches.toList();
-            rawList.sort((o1, o2) -> {
-                HashMap<String, Object> thing1 = (HashMap<String, Object>) o1;
-                HashMap<String, Object> thing2 = (HashMap<String, Object>) o2;
-                return Integer.compare((Integer) thing1.get("match_number"), (Integer) thing2.get("match_number"));
-            });//sort by match number
-            ArrayList<ArrayList<DataPoint>> correctedData = new ArrayList<>();
-            for (Object obj : rawList) {
-                //for each match
-                HashMap<String, Object> match = (HashMap<String, Object>) obj;
-                HashMap<String, Object> matchScoreBreakdown = (HashMap<String, Object>) match.get("score_breakdown");
-                int matchNum = (Integer) match.get("match_number");
-               // System.out.println("Breakdown for match " + matchNum + ": " + matchScoreBreakdown);
-                List<DatabaseManager.MatchRecord> robots = databaseData.stream().filter(matchRecord -> matchRecord.matchNumber() == matchNum).toList();
-                boolean matchComplete = robots.size() == 6;//is the whole match invalid
-                List<DatabaseManager.MatchRecord> redRobots = robots.stream().filter(matchRecord -> matchRecord.position().ordinal() < 3).toList();
-                List<DatabaseManager.MatchRecord> blueRobots = robots.stream().filter(matchRecord -> matchRecord.position().ordinal() > 2).toList();
-                HashMap<String, Object> redAllianceScoreBreakdown  = (HashMap<String, Object>) matchScoreBreakdown.get("red");
-                HashMap<String, Object> blueAllianceScoreBreakdown  = (HashMap<String, Object>) matchScoreBreakdown.get("blue");
-                ArrayList<ArrayList<DataPoint>> correctedMatch = new ArrayList<>();
-                correctedMatch.addAll(correctAlliance(matchComplete, redRobots, redAllianceScoreBreakdown));
-                correctedMatch.addAll(correctAlliance(matchComplete, blueRobots, blueAllianceScoreBreakdown));
-                System.out.println("Corrected Match: " + correctedMatch);
-            }
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
+
+        String eventCode = "2024week0";
+        ArrayList<ArrayList<ArrayList<DataPoint>>> validatedData = DataValidator.validateData(eventCode, databaseData);
     }
 
-    private static ArrayList<ArrayList<DataPoint>> correctAlliance(boolean matchComplete, List<DatabaseManager.MatchRecord> robots, HashMap<String, Object> breakdown) {
-        ArrayList<ArrayList<DataPoint>> correctedAlliance = new ArrayList<>();
-        int autoSpeakerTrue = (int) breakdown.get("autoSpeakerNoteCount");
-        int autoAmpTrue = (int) breakdown.get("autoAmpNoteCount");
-        int teleSpeakerTrue = ((int) breakdown.get("teleopSpeakerNoteAmplifiedCount")) + ((int) breakdown.get("teleopSpeakerNoteCount"));
-        int teleAmpTrue = (int) breakdown.get("teleopAmpNoteCount");
-        for (DatabaseManager.MatchRecord robot : robots) {
-            ArrayList<DataPoint> recordTemp = new ArrayList<>();
-            //for each robot
-            for (DataPoint dataPoint : robot.getDataAsList()) {
-                if (!matchComplete) {
-                    //then there is missing data for at least on robot or there is excess data and the whole match will be marked as incorrect
-                    recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.ErrorLevel.MEDIUM));
-                }
-                try {
-                    switch (Constants.ColumnName.valueOf(dataPoint.getName())) {
 
-                        case TIMESTAMP, MATCH_NUM, TEAM_NUM, ALLIANCE_POS, AUTO_LEAVE, AUTO_COMMENTS, TELE_COMMENTS, ENDGAME_POS -> {
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.ErrorLevel.ZERO));
-                            break;
-                        }
-                        case AUTO_SPEAKER -> {
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.translateErrorNum(robot.autoSpeaker() - autoSpeakerTrue)));
-                            break;
-                        }
-                        case AUTO_AMP -> {
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.translateErrorNum(robot.autoAmp() - autoAmpTrue)));
-                        }
-                        case AUTO_COLLECTED, AUTO_SPEAKER_MISSED, AUTO_AMP_MISSED -> {
-                            //collected - missed = scored + error
-                            //collected - missed - scored = error
-                            int totalMissed = robot.autoAmpMissed() + robot.autoSpeakerMissed();
-                            int totalScored = autoSpeakerTrue + autoAmpTrue;
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.translateErrorNum(robot.autoCollected() - totalScored - totalMissed)));
-                            break;
-                        }
-                        case TELE_SPEAKER -> {
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.translateErrorNum(robot.teleSpeaker() - teleSpeakerTrue)));
-                            break;
-                        }
-                        case TELE_AMP -> {
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.translateErrorNum(robot.teleAmp() - teleAmpTrue)));
-                            break;
-                        }
-                        case TELE_TRAP -> {
-                            int robotPositonNum = (robot.position().ordinal() % 3)+1;
-                            String climb = (String) breakdown.get("endGameRobot" + robotPositonNum);
-                            boolean trap = false;
-                            if (!Objects.equals(climb, "None") && !Objects.equals(climb, "Parked")) {
-                                trap = (boolean) breakdown.get("trap" + climb);
-                            }
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), (trap == (robot.teleTrap() > 0))?(DataPoint.ErrorLevel.ZERO):(DataPoint.ErrorLevel.HIGH)));
-                            break;
-                        }
-                        case TELE_SPEAKER_MISSED, TELE_AMP_MISSED -> {
-                            recordTemp.add(new DataPoint(dataPoint.getName(), dataPoint.getValue(), DataPoint.ErrorLevel.UNKNOWN));
-                            break;
-                        }
-                    }//end switch
-                }catch (IllegalArgumentException messingwithAsher) {
-                    //do nothing this just happens when you try to access the calculated values
-                }
-                System.out.println("Record Temp: " + recordTemp);
-            }
-            correctedAlliance.add(recordTemp);
-
-        }
-        return correctedAlliance;
-
-    }
 
     @FXML
     public void refresh(Event e) {
